+++ stegraca-valise/package.json
@@ -21,6 +21,7 @@
     "react-multi-carousel": "^2.8.4",
     "react-scripts": "^5.0.1",
     "react-sliding-pane": "^7.3.0",
+    "react-textfit": "^1.1.1",
     "sass": "^1.62.1",
     "sqlite3": "^5.1.6",
     "web-vitals": "^2.1.4"
