+++ stegraca-valise/src/components/index.js
@@ -5,5 +5,6 @@ import Header from './header'
 import Details from './details';
 import Footer from './footer';
 import Bio from './biography';
+import Panel from './panel';
 
-export { Header, Valise, Projects, Details, NavBar, Footer, Bio};
+export { Header, Valise, Projects, Details, NavBar, Footer, Bio, Panel};
