+++ stegraca-valise/src/components/navbar/NavBar.jsx
@@ -17,7 +17,7 @@ import { IconContext } from "react-icons";
 export default function NavBar(props) {
   const {navi, setNavi} = props;
 
-  const [isShown, setIsShown] = useState("Contact Me");
+  const [isShown, setIsShown] = useState();
 
   // const handleMouseEnter = (event) => {
   //   let target = setIsShown(event.target.id);
@@ -28,73 +28,61 @@ export default function NavBar(props) {
   return (
     <>
       <div className="nav-block">
-        {/* <div className="link-block">
-            <div className="link-text" onClick={() => setNavi("about")}>
-              About
-            </div>
-        </div> */}
-        <div>
         <div className="nav-title">{isShown}</div>
-          <div className="nav-bar">
+            <ul className="nav-bar">
             <IconContext.Provider value={{ size: "2em" }}>
               <a href="http://www.github.com/omgthegreenranger">
-              <span
+              <li
                 className="nav-btn"
                 id="github"
                 onMouseEnter={() => setIsShown("GitHub")}
                 onMouseLeave={() => setIsShown("")}
               >
                 <SiGithub />
-              </span>
+              </li>
               </a>
               <a href="https://www.linkedin.com/in/stephencardie/">
-              <span
+              <li
                 className="nav-btn"
                 id="linkedin"
                 onMouseEnter={() => setIsShown("LinkedIn")}
                 onMouseLeave={() => setIsShown("")}
               >
                 <SiLinkedin />
-              </span>
+              </li>
               </a>
               <a href="https://mastodon.social/@gogreenranger">
-              <span
+              <li
                 className="nav-btn"
                 id="mastodon"
                 onMouseEnter={() => setIsShown("Mastodon")}
                 onMouseLeave={() => setIsShown("")}
               >
                 <SiMastodon />
-              </span>
+              </li>
               </a>
               <a href="mailto:stephen.cardie@gmail.com">
-              <span
+              <li
                 className="nav-btn"
                 id="email"
                 onMouseEnter={() => setIsShown("Email")}
                 onMouseLeave={() => setIsShown("")}
               >
                 <MdMail />
-              </span>
+              </li>
               </a>
               <a href="#">
-              <span
+              <li
                 className="nav-btn"
                 id="resume"
                 onMouseEnter={() => setIsShown("Resume")}
                 onMouseLeave={() => setIsShown("")}
               >
                 <MdSimCardDownload />
-              </span>
+              </li>
               </a>
             </IconContext.Provider>
-          </div>
-        </div>
-        {/* <div className="link-block">
-          <div className="link-text" onClick={() => setNavi("portfolio")}>
-            Portfolio
-          </div>
-        </div> */}
+            </ul>
       </div>
     </>
   );
