+++ stegraca-valise/src/components/panel/index.js
@@ -1,0 +1,1 @@
+export { default } from './Panel';
