+++ stegraca-valise/src/components/valise/Valise.css
@@ -5,20 +5,23 @@
     width: 100%;
     align-items: center;
     justify-content: center;
+
 }
 
 
-.valise-name {
+.splash-name {
+    display: inline-flex;
     font-weight:bolder;
-    font-size:7em;
-    fill:white;
-    dominant-baseline: hanging;
+    font-size: 11.9cqw;
+    color:white;
+    justify-content: center;
+    line-height: 0.8;
 }
 
 .splash {
     display: flex;
     flex-direction: column;
-    width: 100%;
+    width: 75%;
     align-items: center;
 }
 
@@ -34,14 +37,23 @@
 }
 
 .splash-logo {
+    container-type: inline-size;
     justify-items: center;
-    width:75%;
+    width:100%;
     height:auto;
+}
+
+.splash-logo-top {
+    display: inline-flex;
+    flex-direction: row;
+    width: 100%;
+    justify-content: space-between;
+    position: relative;
 }
 
 .splash-position {
     fill:white;
-    font-size: 2rem;
+    font-size: 3cqw;
     dominant-baseline: hanging;
 }
 
@@ -60,4 +72,13 @@
     align-items:center;
     height: 100%;
     width: 100%;
+    margin:1rem;
+}
+
+.panel {
+    display: flex;
+    flex-direction: column;
+    width: 100%;
+    align-items: center;
+    position:relative;
 }
