+++ stegraca-valise/src/components/valise/Valise.jsx
@@ -1,25 +1,6 @@
 import "./Valise.css";
 import React, { useState } from "react";
-import { Projects, NavBar, Bio } from "../index";
-import {
-    DiCss3,
-    DiHeroku,
-    DiHtml5,
-    DiJsBadge,
-    DiBootstrap,
-    DiMongodb,
-    DiMysql,
-    DiGithub,
-    DiReact,
-  } from "react-icons/di";
-  import {
-    RxCaretLeft,
-    RxCaretRight,
-    RxExternalLink,
-    RxInfoCircled,
-  } from "react-icons/rx";
-  import { IconContext } from "react-icons";
-
+import { Projects, NavBar, Bio, Panel } from "../index";
 export default function Valise(props) {
   // destructure our global states from props
   const [navi, setNavi] = useState("none");
@@ -38,10 +19,12 @@ export default function Valise(props) {
   return (
     <div className="valise-container">
         <div className={`splash ${splashClass}`}>
-            <NavBar navi={navi} setNavi={setNavi} />
+            {/* <NavBar navi={navi} setNavi={setNavi} /> */}
             <Splash navi={navi} setNavi={setNavi} />
         </div>
-      <Panel navi={navi} setNavi={setNavi} />
+        <div className="panel">
+            <Panel navi={navi} setNavi={setNavi} />
+        </div>
     </div>
   );
 }
@@ -49,71 +32,30 @@ export default function Valise(props) {
 function Splash(props) {
     const {navi, setNavi} = props;
   return (
+        // <div className="splash-logo">
+        //   <svg
+        //     width="100%"
+        //     height="100%"
+        //     viewBox="0 0 948 135"
+        //     xmlns="http //www.w3.org/2000/svg"
+        //   >
+        //     <text x="50" y="0" class="splash-position">
+        //       Fullstack Web Developer
+        //     </text>
+        //     <text x="0" y="25" font-width="900px" class="valise-name">
+        //       STEPHEN CARDIE
+        //     </text>
+        //     <line x1="50" y1="130" x2="900" y2="130" stroke="white" />
+        //   </svg>
+        // </div>
         <div className="splash-logo">
-          <svg
-            width="100%"
-            height="100%"
-            viewBox="0 0 948 135"
-            xmlns="http //www.w3.org/2000/svg"
-          >
-            <text x="50" y="0" class="splash-position">
-              Fullstack Web Developer
-            </text>
-            <text x="0" y="25" font-width="900px" class="valise-name">
-              STEPHEN CARDIE
-            </text>
-            <line x1="50" y1="130" x2="900" y2="130" stroke="white" />
-          </svg>
+            <div className="splash-logo-top">
+                <div className="splash-position">Fullstack Web Developer</div>
+                <div><NavBar navi={navi} setNavi={setNavi} /></div>
+            </div>
+            <div>
+                <h1 className="splash-name">STEPHEN CARDIE</h1>
+            </div>
         </div>
-
   );
-}
-
-function Panel(props) {
-    const { navi, setNavi } = props;
-    const handleBack = () => {
-        setNavi("none");
-    }
-    const Display = () => { 
-    switch(navi) {
-        case "about":
-            return(
-                <>
-                <div className="display-panel">
-                <div onClick={handleBack}>Back</div>
-                    <Bio />
-                </div>
-                </>
-            )
-        case "portfolio":
-            return(
-                <>
-                <div className="display-panel">
-                    <div onClick={handleBack}>Back</div>
-                    <Projects /> 
-                </div>
-                </>
-            )
-        default:
-            return(
-                <></>
-            )
-
-    }
-}
-
-
-    return (
-        <>
-        <div className="link-block">
-            <div className="link-text" onClick={() => setNavi("about")}>
-                About
-            </div>
-            <div className="link-text" onClick={() => setNavi("portfolio")}>
-                Portfolio
-            </div>
-        </div>
-            <Display />
-            </>
-    )
 }
